Issue47 comment basis function evaluation (#71)
* reformat lines longer than 120 characters.

* Add a comment in basis_function.h to document the basis function evaluation algorithm.

* comment difference between baf::BasisFunction::IsCoordinateInSupportSpan() and baf::BasisFunction::IsCoordinateInSupport

* Adjust header guards based on google style guide.

* Adjust number of spaces before comments

* Terminate namespaces with comments // namespace "name of namespace"
basef on google style guide

* Terminate namespaces with comments // namespace "name of namespace"
based on google style guide

* small style adjustments based on google style guide.

* delete NOLINT comments.

* Use method AreEqual() from numeric settings while checking difference of knots in knot vector.

* Use method std::equal instead of std::transform and std::any_of.

* try to find reason for warning of undefined behaviour.

* revert changes.

+++ src/elm/element.cc
@@ -27,7 +27,7 @@ int elm::Element::numberOfNodes() const {
 
 double elm::Element::node(int number) const {
 #ifdef DEBUG
-  return nodes_.at(static_cast<unsigned long>(number));
+  return nodes_.at(number);
 #else
   return nodes_[number];
 #endif
